usecase/agi-agent-application/requirements.txt
boto3==1.37.26
botocore==1.37.26
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
langchain==0.3.22
langchain-core==0.3.49
langchain-openai==0.3.11
langchain-teddynote==0.3.45
langgraph==0.3.21
numpy
openai==1.70.0
pydantic==2.11.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
sentence_transformers==4.0.1
tavily_python==0.5.3
tqdm==4.67.1
typing_extensions==4.13.0
Werkzeug==3.1.3
